将Gmv改造成FlinkSQL 支持同时启动Gmv与宽表任务

+++ brain/db/clickhouse/dwd_order_wide.sql
@@ -5,7 +5,7 @@ CREATE TABLE default.dwd_order_wide
 (
     order_id   Int64,
     user_id    Int64,
-    amount     Float64,
+    amount     Decimal(10, 2),
     level      String,
     is_member  Int32,
     ts         Int64
